fix(insert): version-gate add_element_on_transaction (v0=Op::Put, v1=layered)
PR #752 broadened the non-batch insert tree arm so CountSumTree /
ProvableCountTree / ProvableCountSumTree are written as layered subtrees
(Op::PutLayeredReference). grovedb <= v4.1.0 wrote them as plain values
(Op::Put), and that is the behaviour frozen into the live protocol-v11
activation chain (testnet block 245,344, transition_to_version_11). The
layered op computes a different parent value_hash —
combine_hash(value_hash(serialized), NULL_HASH) instead of
value_hash(serialized) — and therefore a different grovedb root, a consensus
divergence on replay.

Rather than revert unconditionally (cf. #757), split add_element_on_transaction
into a versioned dispatch, mirroring the proof v0/v1 pattern:

- v0: grovedb v4.1.0 behaviour — those three types take the Op::Put arm.
  Selected by GROVE_V1 / GROVE_V2, preserving the protocol-v11 root.
- v1: current behaviour — those three types are layered, consistent with the
  batch insert path (both root hash and fee). Selected by GROVE_V3.

v0/v1 live in their own files as frozen snapshots; they differ only in which
match arm those three element types fall into. GROVE_V3's
add_element_on_transaction version slot is bumped 0 -> 1; v1/v2 stay 0.

Adds a consensus-guard test that replays the transition_to_version_11 shape and
pins both roots: the v0 root is byte-identical to PR #757's protocol-v11 golden,
and the v1 (layered) root differs.
