+++ grovedb/src/operations/insert/add_element_on_transaction/v1.rs
@@ -1,0 +1,227 @@
+//! `add_element_on_transaction` — **v1** (current behaviour, `GROVE_V3`+).
+//!
+//! `CountSumTree` / `ProvableCountTree` / `ProvableCountSumTree` are inserted as
+//! **layered subtrees** (`Op::PutLayeredReference`), the same way the batch
+//! insert path writes them. This makes the parent node's `value_hash` =
+//! `combine_hash(value_hash(serialized), child_root_hash)` and the storage cost
+//! the fixed layered tree cost — i.e. the non-batch path agrees with the batch
+//! path on both the root hash and the fee.
+//!
+//! This differs from [`super::v0`] (grovedb v4.1.0 / `GROVE_V1` / `GROVE_V2`)
+//! ONLY in the match arm for those three element types, where v0 uses the
+//! plain-value `Op::Put` path instead. See the module docs in
+//! [`super`][`mod@super`] for the consensus rationale.
+
+use grovedb_costs::{
+    cost_return_on_error, cost_return_on_error_into, cost_return_on_error_no_add, CostResult,
+    CostsExt, OperationCost,
+};
+use grovedb_element::reference_path::path_from_reference_path_type;
+use grovedb_merk::{
+    element::{costs::ElementCostExtensions, insert::ElementInsertToStorageExtensions, ElementExt},
+    tree::NULL_HASH,
+    Merk,
+};
+use grovedb_path::SubtreePath;
+use grovedb_storage::{rocksdb_storage::PrefixedRocksDbTransactionContext, StorageBatch};
+use grovedb_version::version::GroveVersion;
+
+use super::super::InsertOptions;
+use crate::{Element, Error, GroveDb, Transaction};
+
+impl GroveDb {
+    /// `add_element_on_transaction` v1 — see the module documentation.
+    pub(crate) fn add_element_on_transaction_v1<'db, B: AsRef<[u8]>>(
+        &'db self,
+        path: SubtreePath<B>,
+        key: &[u8],
+        element: Element,
+        options: InsertOptions,
+        transaction: &'db Transaction,
+        batch: &'db StorageBatch,
+        grove_version: &GroveVersion,
+    ) -> CostResult<Merk<PrefixedRocksDbTransactionContext<'db>>, Error> {
+        let mut cost = OperationCost::default();
+
+        let mut subtree_to_insert_into = cost_return_on_error!(
+            &mut cost,
+            self.open_transactional_merk_at_path(
+                path.clone(),
+                transaction,
+                Some(batch),
+                grove_version
+            )
+        );
+        // if we don't allow a tree override then we should check
+
+        if options.checks_for_override() {
+            let maybe_element_bytes = cost_return_on_error!(
+                &mut cost,
+                subtree_to_insert_into
+                    .get(
+                        key,
+                        true,
+                        Some(&Element::value_defined_cost_for_serialized_value),
+                        grove_version,
+                    )
+                    .map_err(|e| Error::CorruptedData(e.to_string()))
+            );
+            if let Some(element_bytes) = maybe_element_bytes {
+                if options.validate_insertion_does_not_override {
+                    return Err(Error::OverrideNotAllowed(
+                        "insertion not allowed to override",
+                    ))
+                    .wrap_with_cost(cost);
+                }
+                if options.validate_insertion_does_not_override_tree {
+                    let element = cost_return_on_error_no_add!(
+                        cost,
+                        Element::deserialize(element_bytes.as_slice(), grove_version).map_err(
+                            |_| {
+                                Error::CorruptedData(String::from("unable to deserialize element"))
+                            }
+                        )
+                    );
+                    if element.is_any_tree() {
+                        return Err(Error::OverrideNotAllowed(
+                            "insertion not allowed to override tree",
+                        ))
+                        .wrap_with_cost(cost);
+                    }
+                }
+            }
+        }
+
+        // Dispatch via the underlying element so a NonCounted wrapper takes
+        // the same path as its inner element. The actual `element.insert*`
+        // calls operate on the outer wrapper, which is what we want — the
+        // serialized wrapper bytes go to storage.
+        match element.underlying() {
+            // `ReferenceWithSumItem` shares the reference resolution + proof
+            // shape with `Reference`. The merk feature_type derived from the
+            // element's `sum_value_or_default()` already routes the sum into
+            // any sum-bearing parent; the call site is otherwise identical.
+            Element::Reference(reference_path, ..)
+            | Element::ReferenceWithSumItem(reference_path, ..) => {
+                let path = path.to_vec(); // TODO: need for support for references in path library
+                let reference_path = cost_return_on_error_into!(
+                    &mut cost,
+                    path_from_reference_path_type(reference_path.clone(), &path, Some(key))
+                        .wrap_with_cost(OperationCost::default())
+                );
+
+                let referenced_item = cost_return_on_error!(
+                    &mut cost,
+                    self.follow_reference(
+                        reference_path.as_slice().into(),
+                        false,
+                        Some(transaction),
+                        grove_version
+                    )
+                );
+
+                let referenced_element_value_hash = cost_return_on_error_into!(
+                    &mut cost,
+                    referenced_item.value_hash(grove_version)
+                );
+
+                cost_return_on_error_into!(
+                    &mut cost,
+                    element.insert_reference(
+                        &mut subtree_to_insert_into,
+                        key,
+                        referenced_element_value_hash,
+                        Some(options.as_merk_options()),
+                        grove_version,
+                    )
+                );
+            }
+            // v1: all tree types — including `CountSumTree` / `ProvableCountTree`
+            // / `ProvableCountSumTree` — are written as layered subtrees. This is
+            // the behaviour selected by `GROVE_V3`+; for `GROVE_V1` / `GROVE_V2`
+            // those three types take the plain-value `Op::Put` arm in
+            // [`super::v0`] instead, to preserve the protocol-v11 consensus root.
+            Element::Tree(value, _)
+            | Element::SumTree(value, ..)
+            | Element::BigSumTree(value, ..)
+            | Element::CountTree(value, ..)
+            | Element::CountSumTree(value, ..)
+            | Element::ProvableCountTree(value, ..)
+            | Element::ProvableCountSumTree(value, ..)
+            | Element::ProvableSumTree(value, ..)
+            | Element::ProvableCountProvableSumTree(value, ..) => {
+                if value.is_some() {
+                    return Err(Error::InvalidCodeExecution(
+                        "a tree should be empty at the moment of insertion when not using batches",
+                    ))
+                    .wrap_with_cost(cost);
+                } else {
+                    cost_return_on_error_into!(
+                        &mut cost,
+                        element.insert_subtree(
+                            &mut subtree_to_insert_into,
+                            key,
+                            NULL_HASH,
+                            Some(options.as_merk_options()),
+                            grove_version
+                        )
+                    );
+                }
+            }
+            // CommitmentTree uses BulkAppendTree internally; the initial child
+            // hash must include the empty sinsemilla root so V1 proof
+            // verification works even before the first append.
+            Element::CommitmentTree(..) => {
+                cost_return_on_error_into!(
+                    &mut cost,
+                    element.insert_subtree(
+                        &mut subtree_to_insert_into,
+                        key,
+                        grovedb_commitment_tree::EMPTY_COMMITMENT_TREE_STATE_ROOT,
+                        Some(options.as_merk_options()),
+                        grove_version
+                    )
+                );
+            }
+            // MmrTree, BulkAppendTree, DenseAppendOnlyFixedSizeTree: initial
+            // insert uses NULL_HASH since these trees start empty.
+            Element::MmrTree(..)
+            | Element::BulkAppendTree(..)
+            | Element::DenseAppendOnlyFixedSizeTree(..) => {
+                cost_return_on_error_into!(
+                    &mut cost,
+                    element.insert_subtree(
+                        &mut subtree_to_insert_into,
+                        key,
+                        NULL_HASH,
+                        Some(options.as_merk_options()),
+                        grove_version
+                    )
+                );
+            }
+            Element::Item(..) | Element::SumItem(..) | Element::ItemWithSumItem(..) => {
+                cost_return_on_error_into!(
+                    &mut cost,
+                    element.insert(
+                        &mut subtree_to_insert_into,
+                        key,
+                        Some(options.as_merk_options()),
+                        grove_version
+                    )
+                );
+            }
+            // `underlying()` only unwraps one level; nested wrappers are
+            // forbidden by the constructor and (de)serializer, but the public
+            // insert path can still receive a hand-built nested wrapper —
+            // return a typed error rather than panic.
+            Element::NonCounted(_) | Element::NotSummed(_) | Element::NotCountedOrSummed(_) => {
+                return Err(Error::InvalidInput(
+                    "nested element wrappers are not allowed",
+                ))
+                .wrap_with_cost(cost);
+            }
+        }
+
+        Ok(subtree_to_insert_into).wrap_with_cost(cost)
+    }
+}
